+++ crates/client/sync/src/utils/mod.rs
@@ -1,4 +1,0 @@
-pub mod convert;
-#[cfg(feature = "m")]
-pub mod m;
-pub mod utility;