+++ crates/primitives/chain_config/src/lib.rs
@@ -1,0 +1,5 @@
+mod chain_config;
+mod starknet_version;
+
+pub use chain_config::*;
+pub use starknet_version::*;