+++ crates/primitives/transactions/src/lib.rs
@@ -2,6 +2,7 @@ mod broadcasted_to_blockifier;
 pub mod compute_hash;
 mod from_blockifier;
 mod from_broadcasted_transaction;
+mod from_starknet_core;
 mod from_starknet_provider;
 mod into_starknet_api;
 mod to_starknet_core;
@@ -9,7 +10,7 @@ pub mod utils;
 
 pub use broadcasted_to_blockifier::broadcasted_to_blockifier;
 use dp_convert::ToFelt;
-pub use from_starknet_provider::TransactionTypeError;
+// pub use from_starknet_provider::TransactionTypeError;
 use starknet_api::transaction::TransactionVersion;
 use starknet_types_core::{felt::Felt, hash::StarkHash};