+++ scripts/e2e-coverage.sh
@@ -1,14 +1,25 @@
 #!/bin/bash
 set -e
 
-rm -f target/madara-* lcov.info
+# will also launch anvil and automatically close it down on error or success
 
-source <(cargo llvm-cov show-env --export-prefix)
+anvil --fork-url https://eth.merkle.io --fork-block-number 20395662 &
 
-cargo build --bin deoxys --profile dev
+subshell() {
+    set -e
+    rm -f target/madara-* lcov.info
 
-export COVERAGE_BIN=$(realpath target/debug/deoxys)
-cargo test --profile dev
+    source <(cargo llvm-cov show-env --export-prefix)
 
-cargo llvm-cov report --lcov --output-path lcov.info
-# cargo llvm-cov report
+    cargo build --bin deoxys --profile dev
+
+    export COVERAGE_BIN=$(realpath target/debug/deoxys)
+    cargo test --profile dev
+
+    cargo llvm-cov report --lcov --output-path lcov.info
+    cargo llvm-cov report
+}
+
+(subshell && r=$?) || r=$?
+pkill -P $$
+exit $r